+++ src/main/java/org/klassen/Speisen.java
@@ -5,7 +5,7 @@ public class Speisen extends Artikel {
     //public String beilage;
     final public static int umsatzsteuer = 10;
 
-    public Speisen(String name, int preis/*, String beilage*/) {
+    public Speisen(String name, double preis/*, String beilage*/) {
         this.name = name;
         this.preis = preis;
         //this.beilage = beilage;
